Change types of variables on 104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,17 +6,17 @@
  */
 void fibonacci_98(void)
 {
-unsigned long a, b, c;
+unsigned long long a, b, c;
 int i;
 a = 1;
 b = 2;
-printf("%lu, %lu", a, b);
+printf("%llu, %llu", a, b);
 for (i = 0; i < 96; i++)
 {
 c = a + b;
 a = b;
 b = c;
-printf(", %lu", c);
+printf(", %llu", c);
 }
 printf("\n");
 }
